requirements_pip.txt
absl-py @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/absl-py_1729057589050/work
addict==2.4.0
aiohappyeyeballs @ file:///C:/b/abs_e9fm5mk6ak/croot/aiohappyeyeballs_1734469429350/work
aiohttp @ file:///C:/b/abs_d4_gdjffav/croot/aiohttp_1734687146788/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
attrs @ file:///C:/b/abs_89hmquz5ga/croot/attrs_1734533130810/work
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
cachetools @ file:///C:/b/abs_fcvwpc5vh9/croot/cachetools_1738224687938/work
certifi @ file:///C:/b/abs_43o9lm92p9/croot/certifi_1750863864870/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///C:/b/abs_42rbjngfys/croot/click_1744271606983/work
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1729036581634/work
contourpy @ file:///C:/b/abs_f2u2o_8s9g/croot/contourpy_1732540071787/work
cryptography @ file:///C:/b/abs_ae5lrkywtg/croot/cryptography_1740577842775/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
et_xmlfile @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/et_xmlfile_1729053960370/work
filelock==3.13.1
fonttools @ file:///C:/b/abs_4crkswws2h/croot/fonttools_1737040078745/work
frozenlist @ file:///C:/b/abs_06ctmb1zeo/croot/frozenlist_1730903113463/work
fsspec==2024.6.1
google-auth @ file:///C:/b/abs_86jbgkfwa2/croot/google-auth_1741277194305/work
google-auth-oauthlib==0.4.1
grpcio @ file:///C:/b/abs_98_4ysdoqy/croot/grpc-split_1742490255707/work
h5py @ file:///C:/Users/dev-admin/buildout/croot/h5py_1739563828046/work
idna @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/idna_1729038997316/work
Jinja2==3.1.4
joblib @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/joblib_1729044354027/work
kiwisolver @ file:///C:/b/abs_faf90xet7a/croot/kiwisolver_1737040915779/work
lmdb==1.6.2
Markdown @ file:///C:/b/abs_f9hbi7znfr/croot/markdown_1746114643651/work
MarkupSafe==2.1.5
matplotlib==3.10.0
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730824198951/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730824302834/work
mpmath==1.3.0
multidict @ file:///C:/b/abs_19e3ubo2ew/croot/multidict_1730905504444/work
networkx==3.3
numexpr @ file:///C:/Users/dev-admin/buildout/perseverance-python-buildout/croot/numexpr_1731700337214/work
numpy @ file:///C:/b/abs_56qrev29w_/croot/numpy_and_numpy_base_1738155838326/work/dist/numpy-2.2.2-cp313-cp313-win_amd64.whl#sha256=d740a4ea626e7bf6bf00c35764dc7aa8b81bd2db69d2c126d9f0969e39f92578
oauthlib @ file:///C:/b/abs_5e89byzts4/croot/oauthlib_1750775489850/work
opencv-python==4.10.0
opencv-python-headless==4.10.0
openpyxl @ file:///C:/b/abs_6cdczthntz/croot/openpyxl_1736371239628/work
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas @ file:///C:/b/abs_e18w9ay4we/croot/pandas_1732735115740/work/dist/pandas-2.2.3-cp313-cp313-win_amd64.whl#sha256=89153934c68f42239148879a6c0fd8bb5004e767550db26257dc4551ad0414d7
pathlib @ file:///Users/ktietz/demo/mc3/conda-bld/pathlib_1629713961906/work
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
ply @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/ply_1729041337918/work
propcache @ file:///C:/b/abs_57780xb0gk/croot/propcache_1744012740137/work
protobuf==3.20.3
psutil @ file:///C:/b/abs_b5gv3mn55h/croot/psutil_1736371546320/work
pyasn1 @ file:///C:/Users/dev-admin/buildout/perseverance-python-buildout/croot/pyasn1_1731744867547/work
pyasn1_modules @ file:///C:/b/abs_12mhyss7ds/croot/pyasn1-modules_1752042911584/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///C:/b/abs_6cmmt1gx_z/croot/pyopenssl_1741343840223/work
pyparsing @ file:///C:/Users/dev-admin/buildout/croot/pyparsing_1735861809022/work
PyQt5_sip @ file:///C:/b/abs_d1i1jr9fgt/croot/pyqt-split_1736540551157/work/pyqt_sip
PyQt6==6.7.1
PyQt6_sip @ file:///C:/b/abs_f47pmp6xmn/croot/pyqt-split_1744804513582/work/pyqt_sip
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1729039320841/work
python-dateutil @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/python-dateutil_1729038406576/work
pytz @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pytz_1729039375330/work
PyYAML @ file:///C:/b/abs_79lqlbkifj/croot/pyyaml_1731006239057/work
requests @ file:///C:/b/abs_0b3h9d_724/croot/requests_1750426345668/work
requests-oauthlib @ file:///C:/Users/dev-admin/buildout/perseverance-python-buildout/croot/requests-oauthlib_1731731301130/work
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn @ file:///C:/b/abs_0dwalqot3o/croot/scikit-learn_1737988780693/work/dist/scikit_learn-1.6.1-cp313-cp313-win_amd64.whl#sha256=3e52b503bae0e72bc233231ad5d730537e192e1214e0ac766be348eb9dfb7cfd
scipy @ file:///C:/b/abs_5ejjagtvrz/croot/scipy_1737132284367/work/dist/scipy-1.15.1-cp313-cp313-win_amd64.whl#sha256=fe6c280db2a1d566a65517d2352627e4a6e55f97ca2ca4672e58cecc1fb3469b
setuptools==72.1.0
sip @ file:///C:/b/abs_5cto136kse/croot/sip_1738856220313/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.6.0
threadpoolctl @ file:///C:/b/abs_3904pspv77/croot/threadpoolctl_1731006105881/work
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado @ file:///C:/Users/dev-admin/buildout/croot/tornado_1735843426044/work
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
typing_extensions==4.12.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
Werkzeug @ file:///C:/b/abs_c7bupijx2_/croot/werkzeug_1737448709012/work
wheel==0.45.1
win_inet_pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1729038644686/work
yarl @ file:///C:/b/abs_281qby2vim/croot/yarl_1732546854547/work
